WC2-592 Refactor ETL directory structure
